fix(MOR-881): type float-normalized level fields + cover in conformance
afLevel, rfGain, squelch (ReceiverStatePublic) and powerLevel
(ServerStatePublic) are normalized to float in [0, 1] by
_normalize_public_level_snapshot_value in runtime_helpers.py on the
snapshot path.  Typing them as int caused model_validate to raise
int_from_float errors on real snapshot payloads.

- Changed all four fields to float in state_schema.py.  Pydantic lax
  mode coerces the int(0) default from the dataclass path, so float is
  correct for both producers.
- Added test_snapshot_path_normalized_level_fields_conform: drives the
  snapshot producer with raw int observations (128/200), asserts the
  emitted payload carries float values, and validates the full payload
  against ServerStatePublic.  Confirmed the test fails against old int
  typing with int_from_float.
- Added inline comment for active: Literal["MAIN","SUB"] justifying the
  Literal (only set in three places, always "MAIN" or "SUB").
- TS output unchanged: Python float and int both map to number in JSON
  schema; drift gate passes on committed state.ts.

+++ src/rigplane/web/state_schema.py
@@ -104,9 +104,13 @@ class ReceiverStatePublic(_Strict):
     twinPeakFilter: bool = False
     filterShape: int = 0
     agcTimeConstant: int = 0
-    afLevel: int = 0
-    rfGain: int = 0
-    squelch: int = 0
+    # These three are normalized to float in [0, 1] by the snapshot path
+    # (_normalize_public_level_snapshot_value in runtime_helpers.py).  Pydantic
+    # lax mode coerces the int(0) from the dataclass path, so float is the
+    # correct canonical type for both producers.
+    afLevel: float = 0.0
+    rfGain: float = 0.0
+    squelch: float = 0.0
     sMeter: int = 0
     apfTypeLevel: int = 0
     apfOn: bool = False
@@ -240,10 +244,15 @@ class ServerStatePublic(_Strict):
     updatedAt: str
 
     # Global slow-state / TX flags.
+    # ``active`` is set to "MAIN"/"SUB" in exactly three places:
+    #   _civ_rx.py (0xD2 frame), _dual_rx_runtime.py, and RadioState default.
+    # No other values are produced; the Literal is safe.
     active: Literal["MAIN", "SUB"]
     powerOn: bool = True
     ptt: bool = False
-    powerLevel: int = 0
+    # Normalized to float in [0, 1] by the snapshot path via
+    # _normalize_public_level_snapshot_value (runtime_helpers.py).
+    powerLevel: float = 0.0
     split: bool = False
     dualWatch: bool = False
     scanning: bool = False
